requirements.txt
openai
langchain
Flask
Flask-SQLAlchemy
streamlit
Werkzeug
python-dotenv
azure-storage-blob
tabulate
jsonlines